+++ custom_components/immich_frame/__init__.py
@@ -31,6 +31,7 @@ PLATFORMS = [
     Platform.SELECT,
     Platform.NUMBER,
     Platform.BUTTON,
+    Platform.DATE,
     Platform.SENSOR,
     Platform.SWITCH,
     Platform.TEXT,